Add files via upload
Part 3 – Functions and Classes
1)	Introduction to Scenario
Nowadays, most of the libraries are providing not only physical copies of books but also provide access to eBooks, which have become increasingly popular. Numerous numbers of books in the library need to have a systematic tracking system to store the information about the book, such as title, author, year of publication and others. The customer unique numbers, check-out and return time need to be recorded to track the book’s flow in case anything happens to the book such as it going missing. We can know who needs to be responsible. Librarians sometimes make a mistake when key-in the book’s information, so having the update choice is a good way to keep the library management system running smoothly. 
4)	 Explanation of the Program
This system will utilize a class-based approach by defining the Book class as the parent class and an eBook class as a child class that inherits from Book. The parent class will include attributes such as title, author, and year of publication along with methods for checking out and returning books with the customer’s unique numbers. The child class will extend the parent function by adding specific attributes to eBooks which are file size and download capabilities. The eBook class will override the book info with its own. Then, the update function is available to change the attributes if there are any mistakes during data entry. This structure promotes code reuse and clarity to allow the library to manage its collection effectively. Then, the instances of Book and eBook are created to print their information and demonstrate that the system is functioning correctly.